A shell GUI to contain the entire application in a single window.
Still a work in progress. I've tried several different versions of this but we need something to work on so I am pushing this current version out.
There are a number of issues that concern me which are listed in a comment block starting at line 68.
Currently has the login set as the default composite.